+++ module-apps/application-desktop/windows/Names.hpp
@@ -10,4 +10,5 @@ namespace app::window::name
     inline const std::string desktop_poweroff    = "PowerOffWindow";
     inline const std::string desktop_pin_lock    = "PinLockWindow";
     inline const std::string desktop_update      = "Update";
+    inline const std::string desktop_locked      = "LockedInfoWindow";
 }; // namespace app::window::name