+++ module-cellular/at/src/URC_QIND.cpp
@@ -1,4 +1,4 @@
-#include "../URC_QIND.hpp"
+#include <URC_QIND.hpp>
 #include <log/debug.hpp>
 
 using namespace at::urc;