+++ module-services/service-audio/api/AudioServiceAPI.cpp
@@ -57,18 +57,10 @@ namespace AudioServiceAPI
         return Handle(ret->retCode, ret->token);
     }
 
-    audio::RetCode RoutingRecordCtrl(sys::Service *serv, bool enable)
-    {
-        auto msg    = std::make_shared<AudioRequestMessage>(MessageType::AudioRoutingRecordCtrl);
-        msg->enable = enable;
-
-        return SendAudioRequest(serv, msg)->retCode;
-    }
-
     audio::RetCode RoutingMute(sys::Service *serv, bool enable)
     {
-        auto msg                                 = std::make_shared<AudioRequestMessage>(MessageType::AudioRoutingMute);
-        msg->enable                              = enable;
+        auto msg    = std::make_shared<AudioRequestMessage>(MessageType::AudioRoutingMute);
+        msg->enable = enable;
 
         return SendAudioRequest(serv, msg)->retCode;
     }
@@ -124,11 +116,12 @@ namespace AudioServiceAPI
         return SendAudioRequest(serv, msg)->retCode;
     }
 
+    template <typename T>
     audio::RetCode GetSetting(sys::Service *serv,
                               const audio::Setting &setting,
-                              uint32_t &value,
-                              const audio::Profile::Type &profileType,
-                              const audio::PlaybackType &playbackType)
+                              T &value,
+                              const audio::PlaybackType &playbackType,
+                              const audio::Profile::Type &profileType)
     {
         auto msg  = std::make_shared<AudioGetSetting>(profileType, playbackType, setting);
         auto resp = SendAudioRequest(serv, msg);
@@ -139,21 +132,46 @@ namespace AudioServiceAPI
         return resp->retCode;
     }
 
+    template audio::RetCode GetSetting<uint32_t>(sys::Service *serv,
+                                                 const audio::Setting &setting,
+                                                 uint32_t &value,
+                                                 const audio::PlaybackType &playbackType,
+                                                 const audio::Profile::Type &profileType);
+
+    template audio::RetCode GetSetting<bool>(sys::Service *serv,
+                                             const audio::Setting &setting,
+                                             bool &value,
+                                             const audio::PlaybackType &playbackType,
+                                             const audio::Profile::Type &profileType);
+
+    template <typename T>
     audio::RetCode SetSetting(sys::Service *serv,
                               const audio::Setting &setting,
-                              const uint32_t value,
-                              const audio::Profile::Type &profileType,
-                              const audio::PlaybackType &playbackType)
+                              const T value,
+                              const audio::PlaybackType &playbackType,
+                              const audio::Profile::Type &profileType)
     {
-        auto msg = std::make_shared<AudioSetSetting>(profileType, playbackType, setting, value);
+        auto msg = std::make_shared<AudioSetSetting>(profileType, playbackType, setting, std::to_string(value));
 
         return SendAudioRequest(serv, msg)->retCode;
     }
 
+    template audio::RetCode SetSetting<uint32_t>(sys::Service *serv,
+                                                 const audio::Setting &setting,
+                                                 const uint32_t value,
+                                                 const audio::PlaybackType &playbackType,
+                                                 const audio::Profile::Type &profileType);
+
+    template audio::RetCode SetSetting<bool>(sys::Service *serv,
+                                             const audio::Setting &setting,
+                                             const bool value,
+                                             const audio::PlaybackType &playbackType,
+                                             const audio::Profile::Type &profileType);
+
     std::optional<Tags> GetFileTags(sys::Service *serv, const std::string &fileName)
     {
-        auto msg                                 = std::make_shared<AudioRequestMessage>(MessageType::AudioGetFileTags);
-        msg->fileName                            = fileName;
+        auto msg      = std::make_shared<AudioRequestMessage>(MessageType::AudioGetFileTags);
+        msg->fileName = fileName;
 
         auto resp = SendAudioRequest(serv, msg);
         if (resp->retCode == audio::RetCode::Success) {