+++ module-services/service-audio/api/AudioServiceAPI.hpp
@@ -38,7 +38,6 @@ namespace AudioServiceAPI
      * @return audio::Handle Handle to operation to be used in subsequent operations.
      */
     audio::Handle RoutingStart(sys::Service *serv);
-    audio::RetCode RoutingRecordCtrl(sys::Service *serv, bool enable);
     audio::RetCode RoutingMute(sys::Service *serv, bool enable);
     audio::RetCode RoutingSpeakerPhone(sys::Service *serv, bool enable);
     audio::RetCode RoutingHeadset(sys::Service *serv, bool enable);
@@ -100,11 +99,12 @@ namespace AudioServiceAPI
      * @param playbackType -  type of playback. Not used when profileType is different than playback.
      * @return           Standard service-api return code. Success if suitable.
      */
+    template <typename T>
     audio::RetCode GetSetting(sys::Service *serv,
                               const audio::Setting &setting,
-                              uint32_t &value,
-                              const audio::Profile::Type &profileType = audio::Profile::Type::Idle,
-                              const audio::PlaybackType &playbackType = audio::PlaybackType::None);
+                              T &value,
+                              const audio::PlaybackType &playbackType = audio::PlaybackType::None,
+                              const audio::Profile::Type &profileType = audio::Profile::Type::Idle);
     /** @brief Sets settings. Current profile is taken by default.
      *
      * @param serv - requesting service.
@@ -114,10 +114,11 @@ namespace AudioServiceAPI
      * @param playbackType -  type of playback. Not used when profileType is different than playback.
      * @return           Standard service-api return code. Success if suitable.
      */
+    template <typename T>
     audio::RetCode SetSetting(sys::Service *serv,
                               const audio::Setting &setting,
-                              const uint32_t value,
-                              const audio::Profile::Type &profileType = audio::Profile::Type::Idle,
-                              const audio::PlaybackType &playbackType = audio::PlaybackType::None);
+                              const T value,
+                              const audio::PlaybackType &playbackType = audio::PlaybackType::None,
+                              const audio::Profile::Type &profileType = audio::Profile::Type::Idle);
 
 }; // namespace AudioServiceAPI