+++ module-services/service-cellular/messages/CellularMessage.hpp
@@ -70,7 +70,7 @@ class CellularNotificationMessage : public CellularMessage
         RawCommand,               // send raw command to modem -> returns raw, tokenised result
         PowerDownDeregistering,   // modem informed it has started to disconnect from network
         PowerDownDeregistered,    // modem informed it has disconnected from network
-        NewIncomingUSSD           // modem received new ussd code from network
+        NewIncomingUSSD,          // modem received new ussd code from network
     };
 
     // TODO check and fix all CellularNotificationMessage constructors
@@ -85,7 +85,21 @@ class CellularNotificationMessage : public CellularMessage
     Type type;
     std::string data;
 };
+class CellularTimeNotificationMessage : public CellularMessage
+{
+  private:
+    struct tm time;
 
+  public:
+    CellularTimeNotificationMessage() = delete;
+    explicit CellularTimeNotificationMessage(struct tm time)
+        : CellularMessage(MessageType::CellularTimeUpdated), time(time)
+    {}
+    struct tm getTime(void)
+    {
+        return time;
+    }
+};
 class CellularUSSDMessage : public CellularMessage
 {
   public: