+++ changelog.md
@@ -4,12 +4,15 @@
 
 ### Added
 
+* `[sms]` Added network time synchronisation.
+* `[gui][desktop]` Added *how-to-unlock-screen-info* window.
 * `[gui]` Added GUI timers capability
 * `[fonts]` Added multiplication and division sign to the font files
 * `[bluetooth]` Added storing of BT link keys.
 * `[audio]` AudioMux tests
 * `[audio]` Added vibration logic (effect is visible in logs only)
 * `[settings]` Added DatabaseAgent base class and SettingsAgent.
+* `[audio]` Added vibration/enable sound database logic
 
 
 ### Changed